combine inputs
